requirements.txt
asgiref==3.11.0
Django==4.2
django-cors-headers==4.9.0
djangorestframework==3.14.0
Faker==40.1.0
feedparser==6.0.12
geographiclib==2.1
geopy==2.4.1
psycopg2-binary==2.9.11
python-decouple==3.8
pytz==2025.2
sgmllib3k==1.0.0
sqlparse==0.5.5
tzdata==2025.3
